refactor(utf8_io,utf8_os): compile polyfills on all platforms
