+++ src/api/utf8_io.c
@@ -1,5 +1,12 @@
-#include "../api.h"
-#include "../../utfconv.h"
+#include "api.h"
+
+#ifdef _WIN32
+
+#if (LUA_VERSION_NUM < 503)
+#error "This extension will not work with Lua below version 5.3 or LuaJIT."
+#endif
+
+#include "../utfconv.h"
 
 #include <io.h>
 #include <fcntl.h>
@@ -277,11 +284,10 @@ static const luaL_Reg utf8_iolib[] = {
   { NULL, NULL }
 };
 
-static int utf8_io_enable(lua_State *L) {
-#if (LUA_VERSION_NUM < 503)
-#error "This extension will not work with Lua below version 5.3 or LuaJIT."
 #endif
 
+static int utf8_io_enable(lua_State *L) {
+#ifdef _WIN32
   // here we (attempt to) dynamically detect the version of lua.
   if (lua_version(L) < 503)
     return luaL_error(L, "unsupported runtime Lua version");
@@ -312,12 +318,14 @@ static int utf8_io_enable(lua_State *L) {
     lua_pushcfunction(L, utf8_iolib[i].func);
     lua_settable(L, -3);
   }
+#endif
 
   lua_pushboolean(L, 1);
   return 1;
 }
 
 static int utf8_io_disable(lua_State *L) {
+#ifdef _WIN32
   // check if we've patched the io library
   if (lua_getfield(L, LUA_REGISTRYINDEX, IO_PREFIX "patched") == LUA_TNIL) {
     lua_pushboolean(L, 0);
@@ -339,6 +347,7 @@ static int utf8_io_disable(lua_State *L) {
   // unset patch
   lua_pushnil(L);
   lua_setfield(L, LUA_REGISTRYINDEX, IO_PREFIX "patched");
+#endif
 
   lua_pushboolean(L, 1);
   return 1;
